# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Year of Dates API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type DatePlanDinner = {
  dish: string;
  cuisine: string;
  description: string;
};

export type DatePlanMusic = {
  direction: string;
  artists: string[];
  mood: string;
};

export type DatePlanRitual = {
  title: string;
  description: string;
};

export type DatePlanActivity = {
  title: string;
  description: string;
};

export type DatePlanLocalAddOn = {
  title: string;
  description: string;
};

export interface DatePlan {
  month: number;
  monthName: string;
  theme: string;
  destination: string;
  tagline: string;
  intro: string;
  dinner: DatePlanDinner;
  music: DatePlanMusic;
  ritual: DatePlanRitual;
  conversationPrompts: string[];
  activity: DatePlanActivity;
  localAddOn: DatePlanLocalAddOn;
  effort: string;
  cost: string;
  duration: string;
  scheduledDate?: string | null;
  completed: boolean;
  completedAt?: string | null;
}

export interface UpdateDateBody {
  scheduledDate?: string | null;
  completed?: boolean;
  completedAt?: string | null;
}

export interface Reflection {
  id: number;
  month: number;
  highlight: string;
  memory: string;
  learnedAboutEachOther: string;
  /**
   * @minimum 1
   * @maximum 5
   */
  rating: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreateReflectionBody {
  month: number;
  highlight: string;
  memory: string;
  learnedAboutEachOther: string;
  /**
   * @minimum 1
   * @maximum 5
   */
  rating: number;
}

export interface UpdateReflectionBody {
  highlight?: string;
  memory?: string;
  learnedAboutEachOther?: string;
  /**
   * @minimum 1
   * @maximum 5
   */
  rating?: number;
}

export interface ChecklistItem {
  id: number;
  month: number;
  label: string;
  completed: boolean;
}

export interface ToggleChecklistItemBody {
  completed: boolean;
}

export type FavoriteType = (typeof FavoriteType)[keyof typeof FavoriteType];

export const FavoriteType = {
  ritual: "ritual",
  activity: "activity",
  prompt: "prompt",
} as const;

export interface Favorite {
  id: number;
  type: FavoriteType;
  content: string;
  sourceMonth: number;
  createdAt: string;
}

export type CreateFavoriteBodyType =
  (typeof CreateFavoriteBodyType)[keyof typeof CreateFavoriteBodyType];

export const CreateFavoriteBodyType = {
  ritual: "ritual",
  activity: "activity",
  prompt: "prompt",
} as const;

export interface CreateFavoriteBody {
  type: CreateFavoriteBodyType;
  content: string;
  sourceMonth: number;
}

export type LearningAbout = (typeof LearningAbout)[keyof typeof LearningAbout];

export const LearningAbout = {
  seth: "seth",
  elana: "elana",
  both: "both",
} as const;

export interface Learning {
  id: number;
  month: number;
  about: LearningAbout;
  content: string;
  createdAt: string;
}

export type CreateLearningBodyAbout =
  (typeof CreateLearningBodyAbout)[keyof typeof CreateLearningBodyAbout];

export const CreateLearningBodyAbout = {
  seth: "seth",
  elana: "elana",
  both: "both",
} as const;

export interface CreateLearningBody {
  month: number;
  about: CreateLearningBodyAbout;
  content: string;
}

export interface Summary {
  totalDates: number;
  completedDates: number;
  upcomingMonth?: number | null;
  totalReflections: number;
  totalLearnings: number;
  nextScheduledDate?: string | null;
}