# lib/api-zod/src/generated/types/checklistItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Year of Dates API
 * OpenAPI spec version: 0.1.0
 */

export interface ChecklistItem {
  id: number;
  month: number;
  label: string;
  completed: boolean;
}