# lib/api-zod/src/generated/types/createFavoriteBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Year of Dates API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateFavoriteBodyType } from "./createFavoriteBodyType";

export interface CreateFavoriteBody {
  type: CreateFavoriteBodyType;
  content: string;
  sourceMonth: number;
}