# lib/api-zod/src/generated/types/createLearningBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Year of Dates API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateLearningBodyAbout } from "./createLearningBodyAbout";

export interface CreateLearningBody {
  month: number;
  about: CreateLearningBodyAbout;
  content: string;
}