# lib/api-zod/src/generated/types/createReflectionBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Year of Dates API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateReflectionBody {
  month: number;
  highlight: string;
  memory: string;
  learnedAboutEachOther: string;
  /**
   * @minimum 1
   * @maximum 5
   */
  rating: number;
}