# lib/api-zod/src/generated/types/learning.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Year of Dates API
 * OpenAPI spec version: 0.1.0
 */
import type { LearningAbout } from "./learningAbout";

export interface Learning {
  id: number;
  month: number;
  about: LearningAbout;
  content: string;
  createdAt: string;
}