# lib/api-zod/src/generated/types/updateDateBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Year of Dates API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateDateBody {
  scheduledDate?: string | null;
  completed?: boolean;
  completedAt?: string | null;
}